instructions/requirements.txt
psycopg2-binary==2.9.3
dash
dash-daq
pandas
plotly
gunicorn
datetime
